Default environment variables for launching the shell
Render properly the topline which I had not noticed, but was
off-location.

Use a swift-ish idiom for loops, to prevent future errors, so
I use "defer { POST }" after a while loop, due to Swift lack of
for loops with the operation-at-end.

Charsets - update for Danish character to use "ĳ" unicode character
rather than the two separate ascii values ij.   This solves a long-standing
problem in XtermSharp.

Handful of bug fixes on the Escape sequence parser, the UInt8 math
meant that overflow a UInt8 produced a zero, not preserved the value
beyond UInt8, so the tables were wrong.

Cursor now updates the position always, regardless of the update region

Cursor now shows the proper color

Better, but not perfect byte buffer to character renderer.