Finally, both dictation and international input mechanism are working.
There is still a pending crash, with chinese input, I seem to be
creating an invalid text range, and can crash, to reproduce: dictate
"hello world", then enter some multi-character chinese characters,
then press delete.  This creates an invalid range that ends up
requesting an invalid range of text.

A defensive check would work, but we should get to the root of it, so
I am keeping the abort for now.