Restructured a bit how the Sixel image loading will take place, rather than embedding platform specific information inside Terminal (AppKit/UIKit), the handling of the image is delegated to the platform specific part.
Made some of the delegate methods of Terminal follow the same convention
than the rest already had.

Started to think about how I am going to render the sixel data.   It
looks like the image is in the wrong place, and it is too large, I think
I might need to scale it.

Also, currently this has a side-effect, the buildAttributedString is poking
at a global `images` array, which I do not like.  I need a better system
to render, but I want to avoid walking the terminal buffer twice, so a
variation of this is still needed.

The code is currently disabled.