Fixes the scrolling behavior, it was previously snapping back to the original location, this was caused by the hit variable in the .changed event for the pan handler not being updated with the terminal buffer offset, like it was done on the .began section.
This uncovered a misuse of the Pan API as well, where I was updating
the panStart on every change that took place, which produced an
unpleasant jumpy behavior.    The fix is to not update the panStart location
and rely entirely on the wisdom of the gestureRecognizer to give us the
proper information.

Lastly, this fixes another crash with the buffer
updateSelectionAttributesIfNeeded.  At the end of the day, this can be
caused by a selection that is out of sync (I fixed a bug previously,
but did not properly document on the commit, so it is hard to track
down, but at the time, it boiled down to having to turn off the
selection after some changes).  In this particular case, a font change
will change the dimensions of the buffer, so a call to
fullBufferUpdate(terminal:) is required to keep the data structures in
sync.