This is the proper fix for the double-sized characters issue that we have struggled with for a long time.   It shows up with both emoji and chinese characters - and it is rooted on the fact that were attempting to track the current column based on the glyphCount on a run - and this is wrong for too many reasons.  A glyph could use two spaces and not just one, but a glyph could also contain two characters linked instead of one.
The terminal internally tracks the column based on the character width that we
compute - which is an estimation (see previous TODOs) based on known characters
sizes (which btw, have changed over history, so this is not a precise science).

Instead, what this code does now is to return segments for a line grouped by the
character width and their column start.   Then the screen renderer can match the
column with the strict grid column to start rendering at that position.

This undoes the previous hack that added an empty space after a double-width
character.