This has bothered me for too long, and the previous patch sort of made me assume "With this new segmenting, the column starts are correct, and the CTRuns will at least start in the right location, so hopefully text from other languages should work out of the box".
And then I tried arabic, and it looked horrible.  And I was reminded of this
very old patch https://github.com/migueldeicaza/SwiftTerm/issues/227 which we
had never finished - because at the time I did not realize that we had an
impedance mismatch mismatch between the column tracking and the glyphCount that
we were using.   With that out of the way, this renders properly arabic.

That said, I am afraid that selection is still off.